move joinAllPaths and map_pathinfo to contourcarpet/plot.js
... no need to have separate file for these routines.
